pyproject.toml
[project]
name = "mcp-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-aws>=0.2.23",
    "langchain-community>=0.3.24",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "mcp[cli]>=1.8.1",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "paramiko>=3.5.1",
    "python-dotenv>=1.1.0",
]
